# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API HỆ THỐNG CỬA HÀNG ĐIỆN THOẠI PHÚC MOBILE",
        "description": "Hệ thống API quản lý cửa hàng kinh doanh điện thoại di động và phụ kiện công nghệ PHÚC MOBILE. Hỗ trợ đầy đủ phân quyền người dùng (RBAC) và chuẩn hóa dữ liệu quan hệ 3NF.\n *     \n *     ### Danh sách Vai trò (Roles):\n *     - **superadmin**: Quyền cao nhất, quản lý toàn bộ hệ thống bao gồm cả nhân viên và khách hàng.\n *     - **admin**: Quản lý nghiệp vụ chính (Điện thoại, Hãng sản xuất, Danh mục, Mã giảm giá, Đơn hàng, Đánh giá).\n *     - **staff**: Nhân viên vận hành cửa hàng, xử lý trạng thái đơn hàng và phê duyệt đánh giá sản phẩm.\n *     - **user**: Khách hàng (Xem sản phẩm, thêm giỏ hàng, đặt mua điện thoại và gửi đánh giá).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server (Tự động nhận diện Host)"
        }
    ],
    "paths": {
...
        "/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Đăng ký tài khoản",
                "operationId": "0903f2e0df1a8857bee538fc055417f0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Nguyen Van A"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "user@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123456"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Đăng ký thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Đăng ký thành công"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "eyJhbGciOiJIUz..."
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Lỗi validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Dữ liệu gửi lên không hợp lệ."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "email": [
                                                    "Email đã được sử dụng."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Đăng xuất",
                "operationId": "68a6108af04af5e3bc1ea68a6c6f1299",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Đăng xuất thành công"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Lấy thông tin người dùng đang đăng nhập",
                "operationId": "e18baa7d6dfc432241dca2fc1985ce40",
                "responses": {
                    "200": {
                        "description": "Thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/forgot-password": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Quên mật khẩu - gửi link đặt lại qua email",
                "operationId": "da27e3923ff357f8610f31cf2cb83224",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "user@gmail.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Gửi email thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Link đặt lại mật khẩu đã được gửi vào email của bạn."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation lỗi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "errors": {
                                            "type": "string",
                                            "example": "Email không tồn tại trong hệ thống."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/auth/google/callback": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Xử lý callback từ Google (Hệ thống tự động gọi)",
                "description": "HÀNH VI TỰ ĐỘNG: Endpoint này được thiết kế để Google gọi lại sau khi người dùng xác thực thành công. Mã 'code' được Google cấp tự động và chỉ có hiệu lực một lần trong vài giây. Không nên gọi endpoint này một cách thủ công từ Swagger/Postman.",
                "operationId": "0da573308ea6982cdd61fa5120d9e835",
                "parameters": [
                    {
                        "name": "code",
                        "in": "query",
                        "description": "Mã xác thực một lần (One-time code) do Google cấp",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirect về Frontend kèm token hoặc lỗi"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "User": {
                "title": "User",
                "description": "User model schema",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Nguyen Van A"
                    },
                    "email": {
                        "type": "string",
                        "example": "user@gmail.com"
                    },
                    "phone": {
                        "type": "string",
                        "example": "0123456789",
                        "nullable": true
                    },
                    "gender": {
                        "type": "string",
                        "enum": [
                            "male",
                            "female",
                            "other"
                        ],
                        "example": "male",
                        "nullable": true
                    },
                    "birthday": {
                        "type": "string",
                        "format": "date",
                        "example": "1990-01-01",
                        "nullable": true
                    },
                    "avatar": {
                        "type": "string",
                        "example": "avatar.png",
                        "nullable": true
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Nhập JWT token vào đây để thực hiện các yêu cầu có bảo mật.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}